Initial refactoring to abstract engine interface
