+++ pkg/engine/builtin/engine_test.go
@@ -1,4 +1,4 @@
-package engine
+package builtin
 
 import (
 	"fmt"
